pyproject.toml
[tool.poetry]
name = "nirs"
version = "0.1.0"
description = ""
authors = ["Psycho Asto <kuvms068@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
Django = "^5.0.3"
aiogram = "^3.4.1"
django-filter = "^24.1"
redis = "^5.0.3"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.3.1"
django-debug-toolbar = "^4.3.0"
drf-spectacular = "^0.27.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
